# lib/api-zod/src/generated/types/reportUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DataLens — AI-powered business intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportUpdateType } from './reportUpdateType';

export interface ReportUpdate {
  title?: string;
  description?: string;
  type?: ReportUpdateType;
  category?: string;
  data?: string;
}